refactoring
Moving some of the time calculation logic into separate functions.

I’m a bit confused here on if a new test is needed, but it seems like
it shouldn’t be since I’m not really adding new code and I have all the
existing tests as proof that it still works.